Moves fragments and activities into directory for readibility, built a ViewModel to continue the Int Id of the last Navigation Icon hit, my hope is that at some point I will be able to read that int in my FrameLayout Fragment and select the appropriate fragment to display
