+++ app/src/main/java/com/michaeljahns/namespace/UIViewModel.kt
@@ -1,0 +1,11 @@
+package com.michaeljahns.namespace
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class UIViewModel : ViewModel() {
+    var intView = MutableLiveData<Int>()
+    fun select(int: Int) {
+        intView.value = int
+    }
+}
