+++ app/src/main/java/com/michaeljahns/namespace/activities/MainActivity2.kt
@@ -1,4 +1,4 @@
-package com.michaeljahns.namespace
+package com.michaeljahns.namespace.activities
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
