+++ app/src/main/res/layout/fragment_frame_layout.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FrameLayout">
+    tools:context=".fragments.FrameLayout">
 
 </FrameLayout>
