+++ app/src/main/res/layout/fragment_navigation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:Context=".NavbarFragment">
+    tools:Context=".NavigationFragment">
 
     <com.google.android.material.bottomappbar.BottomAppBar
         android:id="@+id/mainAppBar"
