+++ app/src/main/res/layout/fragment_scenario.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScenarioFragment">
+    tools:context=".fragments.ScenarioFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/vp2Scenario"
